make generic AreayIndex, use it in way and route indexes
